fix: InterviewFInalRepository value Long으로 수정

+++ src/main/java/com/tave/tavewebsite/domain/interviewfinal/repository/InterviewFinalRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface InterviewFinalRepository extends JpaRepository<InterviewFinal, Integer> {
+public interface InterviewFinalRepository extends JpaRepository<InterviewFinal, Long> {
 
     Page<InterviewFinal> findAllByOrderByInterviewDateAscInterviewTimeAscUsernameAsc(Pageable pageable);
